fix: rename use-wakelock hook
